Feat: Add local user security

+++ application.py
@@ -130,20 +130,21 @@ def gconnect():
 @app.route('/categories')
 def showCategories():
     '''Shows all categories associated with a user in the database'''
-    # edit to add user logic
-    categories = session.query(Category).all()
-    return render_template("categories.html", categories = categories)
-    # return "Categories page"
+    if 'username' not in login_session:
+        return redirect('/login')
+    else:
+        categories = session.query(Category).filter_by(user_id=login_session['user_id']).all()
+        return render_template("categories.html", categories = categories)
 
 # Create new category
 @app.route('/category/new', methods = ['GET', 'POST'])
 def newCategory():
-    # if 'username' not in login_session:
-    #    return redirect('/login')
+    if 'username' not in login_session:
+       return redirect('/login')
     if request.method == 'POST':
         newCategory = Category(
             name=request.form['name'],
-            # add user_name here
+            user_id=login_session['user_id']
             )
         session.add(newCategory)
         flash('New Category %s Successfully Created' % newCategory.name)
@@ -157,10 +158,10 @@ def newCategory():
 def editCategory(category_id):
     editedCategory = session.query(
     	Category).filter_by(id=category_id).one()
-    # if 'username' not in login_session:
-    #     return redirect('/login')
-    # if editedCategory.user_id != login_session['user_id']:
-    #     return "<script>function myFunction() {alert('You are not authorized to edit this category. Please create your own category in order to edit.');}</script><body onload='myFunction()'>"
+    if 'username' not in login_session:
+        return redirect('/login')
+    if editedCategory.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this category. Please create your own category in order to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             editedCategory.name = request.form['name']
@@ -174,10 +175,10 @@ def editCategory(category_id):
 def deleteCategory(category_id):
     categoryToDelete = session.query(
     	Category).filter_by(id=category_id).one()
-    # if 'username' not in login_session:
-    #     return redirect('/login')
-    # if restaurantToDelete.user_id != login_session['user_id']:
-    #     return "<script>function myFunction() {alert('You are not authorized to delete this category. Please create your own category in order to delete.');}</script><body onload='myFunction()'>"
+    if 'username' not in login_session:
+        return redirect('/login')
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this category. Please create your own category in order to delete.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(categoryToDelete)
         session.query(Item).filter_by(category_id=category_id).delete()
@@ -191,29 +192,28 @@ def deleteCategory(category_id):
 @app.route('/category/<int:category_id>')
 def showCategory(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
-    # creator = getUserInfo(category.user_id)
+    creator = getUserInfo(category.user_id)
     items = session.query(Item).filter_by(
         category_id=category_id).all()
-    # if 'username' not in login_session or creator.id != login_session['user_id']:
-    #     return render_template('publicmenu.html', items=items, restaurant=restaurant, creator=creator)
-    # else:
-    #    return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
-    return render_template('category.html', items=items, category=category) # , creator=creator)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return redirect('/login')
+    else:
+        return render_template('category.html', items=items, category=category, creator=creator)
 
 # Create new item in a category
 @app.route('/category/<int:category_id>/item/new', methods = ['GET', 'POST'])
 def newCategoryItem(category_id):
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     category = session.query(Category).filter_by(id=category_id).one()
-    # if login_session['user_id'] != category.user_id:
-    #     return "<script>function myFunction() {alert('You
-    #     are not authorized to add items to this category.
-    #     Please create your own category in order to add items.');}</script><body onload='myFunction()'>"
+    if login_session['user_id'] != category.user_id:
+        return """<script>function myFunction() {alert('You
+        are not authorized to add items to this category.
+        Please create your own category in order to add items.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         newItem = Item(name=request.form['name'],
                 description=request.form['description'],
-                category_id=category_id) # , user_id=category.user_id)
+                category_id=category_id, user_id=category.user_id)
         session.add(newItem)
         session.commit()
         flash('New Item,  %s, Successfully Created' % (newItem.name))
@@ -224,14 +224,14 @@ def newCategoryItem(category_id):
 # Edit item in a category
 @app.route('/category/<int:category_id>/item/<int:item_id>/edit', methods = ['GET', 'POST'])
 def editItem(category_id, item_id):
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     editedItem = session.query(Item).filter_by(id=item_id).one()
     category = session.query(Category).filter_by(id=category_id).one()
-    # if login_session['user_id'] != categor.user_id:
-    #     return "<script>function myFunction() {alert('You
-    #     are not authorized to edit items in this category.
-    #     Please create your own category in order to edit items.');}</script><body onload='myFunction()'>"
+    if login_session['user_id'] != categor.user_id:
+        return """<script>function myFunction() {alert('You
+        are not authorized to edit items in this category.
+        Please create your own category in order to edit items.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
